remove Debian packaging material from this branch
I ask for Signed-off-by tags which means that if pkg-perl team members
commit to debian/, I might not be able to pull that work into my
master branch.  Then I'd end up with an upstream master with old
debian/ and salsa master with new debian/ which is undesirable.

So, I'll work on debian/ on a separate branch which won't require
Signed-off-by tags for every commit.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>